refactor: drop dead WorkspaceStore.Update and AppendOpenTab
These two exported, flock-coordinated methods had no non-test, non-interface caller: they are absent from the app's WorkspaceStore interface, and production persists open tabs by overwriting the whole workspace via workspaceService.Save (app_persistence.go), never through this atomic read-modify-write path. Remove both methods and their tests, relocating the still-live saveWorkspaceLocked (called by the discovery merge) into workspace_store.go next to Save so the tabs file can be deleted. Rewire the UpsertFromDiscovery lost-update concurrency test to drive its tab writer through the production load+Save append path instead of the deleted AppendOpenTab. Part of the quality stacked diff (stacked on roadmap/03-applytmuxenvfromconfig-force-true-branch-and-s).
