--- internal/data/workspace_store_tabs.go
@@ -1,93 +1,0 @@
-package data
-
-import (
-	"encoding/json"
-	"errors"
-	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/andyrewlee/amux/internal/fsatomic"
-)
-
-// Update atomically loads a workspace, passes it to fn for modification, and
-// saves it back. The workspace flock is held across the entire Load+fn+Save
-// sequence, preventing lost-update races between concurrent amux processes.
-func (s *WorkspaceStore) Update(id WorkspaceID, fn func(ws *Workspace) error) error {
-	if err := validateWorkspaceID(id); err != nil {
-		return err
-	}
-
-	lockFiles, err := s.lockWorkspaceIDs(id)
-	if err != nil {
-		return err
-	}
-	defer unlockRegistryFiles(lockFiles)
-
-	ws, err := s.load(id, true)
-	if err != nil {
-		return err
-	}
-	ws.storeID = id
-
-	if err := fn(ws); err != nil {
-		return err
-	}
-
-	return s.saveWorkspaceLocked(id, ws)
-}
-
-// AppendOpenTab atomically appends a tab to workspace metadata.
-// It acquires the workspace lock, reloads current metadata, and writes once to
-// avoid lost-tab updates when multiple agent runs complete concurrently.
-func (s *WorkspaceStore) AppendOpenTab(id WorkspaceID, tab TabInfo) error {
-	if err := validateWorkspaceID(id); err != nil {
-		return err
-	}
-
-	lockFiles, err := s.lockWorkspaceIDs(id)
-	if err != nil {
-		return err
-	}
-	defer unlockRegistryFiles(lockFiles)
-
-	ws, err := s.load(id, true)
-	if err != nil {
-		return err
-	}
-	ws.storeID = id
-
-	sessionName := strings.TrimSpace(tab.SessionName)
-	if sessionName != "" {
-		for _, existing := range ws.OpenTabs {
-			if strings.TrimSpace(existing.SessionName) == sessionName {
-				return nil
-			}
-		}
-	}
-	ws.OpenTabs = append(ws.OpenTabs, tab)
-
-	return s.saveWorkspaceLocked(id, ws)
-}
-
-func (s *WorkspaceStore) saveWorkspaceLocked(id WorkspaceID, ws *Workspace) error {
-	if ws == nil {
-		return errors.New("workspace is required")
-	}
-	path := s.workspacePath(id)
-	dir := filepath.Dir(path)
-	if err := os.MkdirAll(dir, 0o755); err != nil {
-		return err
-	}
-	data, err := json.MarshalIndent(ws, "", "  ")
-	if err != nil {
-		return err
-	}
-	// Atomic replace (temp + fsync + rename, with backup recovery on platforms
-	// that need it), matching Save. The caller already holds the workspace lock.
-	// A crash mid-save can never leave a truncated workspace.json behind.
-	if err := fsatomic.WriteFile(path, data, 0o644); err != nil {
-		return err
-	}
-	return nil
-}
